pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md" 

packages = [
    { include = "hackathon" },
]

[tool.poetry.dependencies]
python = ">=3.10.13,<3.11"
matplotlib = "^3.8.2"
plotly = "^5.18.0"
streamlit = "^1.28.2"
python-dotenv = "^1.0.0"
langchain = "^0.0.343"
huggingface-hub = "^0.19.4"
uvicorn = "^0.24.0.post1"
sse-starlette = "^1.8.2"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
starlette-context = "^0.3.6"
streamlit-chat = "^0.1.1"
nbformat = "^5.9.2"
opensearch-py = "^2.4.2"
boto3 = "^1.34.14"
folium = "^0.15.1"
streamlit-folium = "^0.18.0"
geopandas = "^0.14.3"
streamlit-gov-uk-components = "^0.0.14"
fuzzywuzzy = "^0.18.0"
botocore = "^1.34.64"
instructor = "^1.2.0"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
llama-cpp-python = "^0.2.52"
ipykernel = "^6.29.3"
jupyter = "^1.0.0"
sentence-transformers = "^2.4.0"
pre-commit = "^3.7.0"
black = {extras = ["jupyter"], version = "^24.3.0"}
nbstripout = "^0.7.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
